[misp-feed] Add: Check the json of an event is a list or not.
A misp-feed service which we use returns abnormal format json such as `[{"Event": ...}]` as an Event.
I added the check and get one row in the json mechanism to import such feeds.

Of course, this json file maybe include one row in this case.
I also added the length check of it.