[crits] Ingest non-report relationships, and reduce duplicates
This change brings in relationships that are present in CRITs, but
between objects that don't share a report - I call these "global"
relationships. Another change here is to de-duplicate relationships:
CRITs records two relationships for each relationship between objects,
a From->To in and a corresponding, but reversed, copy of the same
relationship in the second/dest entity. These will now only create one
side of the relationship in OpenCTI (which automatically implies the
reverse). The presence of this pairing in CRITs is mainly due to
limitations of its object model on a non-relational database.