[crits] Create the connector (#849)
* [crits] Add initial code for CRITs connector

This is intended to be a connector that will fetch data from an external
CRITs (https://crits.github.io) instance, and import it into OpenCTI.
Similar to the existing "external-import/misp" connector.

* [crits] Start ingesting IPs, Domains, and Events

The Domains and IPs that are related to Events will be imported as
contents of those reports (Events).

* [crits] observe update_existing_data connector config

* [crits] Ingest Reports+contents first, then all collections

Some CRITs object types don't store "source" information inside the
object. So, ingest Events->Reports, plus their first-degree relations as
contents, first. This enables the connector to use the source
information for "author" from the parent Report for any initial ingest,
reducing the number of entites created in OpenCTI with no author despite
a reliable authorship link defined in CRITs.

* [crits] Do not assume all entites have a source list

* [crits] Ingest Campaigns as either IntrusionSet or Campaign

Added a new config parameter that controls how CRITs Campaigns are
ingested into OpenCTI. In my experience, this collection is typically
used to house IntrusionSet information, so I chose that as the default,
but anyone may change it to Campaign if they prefer.

* [crits] Import Actors as ThreatActor entities

* [crits] Ingest CRITs "Backdoors" as Malware entities

* [crits] Import Exploit data as Malware w/ exploit-kit subtype

* [crits] Import bucket_list items as OpenCTI labels

* [crits] Comment out CVE assignment until ready for use

* [crits] Initial work to ingest Indicator types

Just works on IPs and Domains, and doesn't yet forge relationships with
linked TLOs in CRITs that are imported as Observables

* [crits] Ingest Sample objects as "File" Observables

This brings in the metadata that maps to the STIX2.1 File objects, but
does not yet import the artifact content

* [crits] Add external reference to the CRITs page for imported data

* [crits] Import Emails as "EmailMessage" observables

This isn't bringing over the contents as artifacts yet, but populates
the STIX metadata fields

* [crits] Ingest signatures as appropriate pattern-type indicators

* [crits] Ingest Target objects as Identities of type "individual"

* [crits] Import RawData (small Text and JSON artifacts) as Text observables

* [crits] Consolidate bundles into a single work_id each run

* [crits] Upload artifacts if filedata is present

* [crits] Use "contains_refs" instead of unsupported(?) "content_ref"

* [crits] Fix error message to properly report events collection failure

* [crits] Allow artifacts created from samples to relate to report

* [crits] Import relationships that are part of an event

When two items sharing a CRITs relationship are found in a single
report, then store the relationship as part of the report in OpenCTI.

* [crits] Ingest non-report relationships, and reduce duplicates

This change brings in relationships that are present in CRITs, but
between objects that don't share a report - I call these "global"
relationships. Another change here is to de-duplicate relationships:
CRITs records two relationships for each relationship between objects,
a From->To in and a corresponding, but reversed, copy of the same
relationship in the second/dest entity. These will now only create one
side of the relationship in OpenCTI (which automatically implies the
reverse). The presence of this pairing in CRITs is mainly due to
limitations of its object model on a non-relational database.

* [crits] Propagate the default marking definition to more places

* [crits] Make timestamp field, chunk size, score, and TLP configurable

* [crits] Fix log_warn -> log_warning

* [crits] Fix misspelling GREEM -> GREEN

* [crits] Update docker-compose.yml with new configuration vars

* [crits] Fix two misspellings of configuration variables

* [crits] Fix divide-by-zero when no data to consume

* [crits] Support the connect-run-and-terminate method

* [crits] Manage last-run time in connector state

By default, the connector will try to consume all of the data, and then
it will store the current run-time in the connector state. On next run
attempt, it will only consume data that has a newer timestamp (based
upon crits.timestamp_field) than the last stored run time

* [crits] Fix formatting

* [crits] Add more documentation in the behavior section

* [crits] Update to 5.3.16 to match current OpenCTI version

* [crits] Refactor indicator template to be more versatile

* [crits] Support more atomic Indicator types from CRITs