[shodan-internetdb] Limit STIX object query to needed fields
This reduces the amount of wasted data transfer, and also the prior
behavior was including the list of imports, which would cause a request
to S3 or Minio for LIST every time the read was called. In addition to
slowing down the work, it also would increase S3 costs in AWS because
LIST is an expensive AWS operation (10x more than GET).