[yara] Use customAttributes to limit overhead of yara sig list() call
Since an Indicator can have importFiles associated with it, when the
list call is performed to get all Indicators that match the
pattern_type="yara" filter, it also attempts to retrieve any importFiles
or exportFiles (by performing an S3 LIST operation), but the results of
this will never be used. In fact, the yara connector only cares about
id, standard_id, pattern, name, and pattern_type. This change limits the
request to those fields, and never attempts a S3 LIST operation. This
drastically speeds up execution, as well as significantly reduces
monetary charges when using AWS S3 as the file storage backend.