[all] Set CONNECTOR_TYPE in Dockerfiles
It must be set to a **static** value, remove it from user-facing config.
The knob is unneeded noise, and a place to introduce errors.

Existing installations, ie setting the env in compose, are fine.

Removed README.md documentation, as it was empty and meaningless.

README.md#Assumptions: Docker is used for Docker and deployment.
  For non-Docker development: config.yml is not touched in commit.