[eset] Manually recompute the STIX id for malware entity before ingesting
The STIX2 malware entities from ESET's data stream appear to have their
IDs randomly generated each time they appear in a new object. This
results in many thousands of STIX IDs being generated for the same
malware signature. To correct this, the STIX ID provided by ESET is
overridden by the deterministic id generator available in pycti, before
the entity is ingested. This is contributing to redis memory
consumption, as well as unnecessary backlogging of the ESET connector's
ingest queue.